requirements.txt
Django==3.2
pymysql==1.0.2
django-add-default-value==0.10.0
django-migration-linter==4.0.0
google-api-python-client==2.44.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
psycopg2-binary==2.9.3
django-heroku==0.0.0
gunicorn==20.1.0
minio==7.1.7
djangorestframework==3.13.1
